+++ data/src/main/java/com/threegap/bitnagil/data/auth/datasourceimpl/AuthRemoteDataSourceImpl.kt
@@ -6,16 +6,18 @@ import com.threegap.bitnagil.data.auth.model.request.TermsAgreementRequest
 import com.threegap.bitnagil.data.auth.model.request.WithdrawalReasonRequest
 import com.threegap.bitnagil.data.auth.model.response.LoginResponse
 import com.threegap.bitnagil.data.auth.service.AuthService
+import com.threegap.bitnagil.data.auth.service.LoginService
 import com.threegap.bitnagil.data.common.safeApiCall
 import com.threegap.bitnagil.data.common.safeUnitApiCall
 import javax.inject.Inject
 
 class AuthRemoteDataSourceImpl @Inject constructor(
     private val authService: AuthService,
+    private val loginService: LoginService,
 ) : AuthRemoteDataSource {
     override suspend fun login(socialAccessToken: String, loginRequest: LoginRequest): Result<LoginResponse> =
         safeApiCall {
-            authService.postLogin(socialAccessToken, loginRequest)
+            loginService.postLogin(socialAccessToken, loginRequest)
         }
 
     override suspend fun submitAgreement(termsAgreementRequest: TermsAgreementRequest): Result<Unit> =
@@ -35,6 +37,6 @@ class AuthRemoteDataSourceImpl @Inject constructor(
 
     override suspend fun reissueToken(refreshToken: String): Result<LoginResponse> =
         safeApiCall {
-            authService.postReissueToken(refreshToken)
+            loginService.postReissueToken(refreshToken)
         }
 }
